Log & compatibility update
Logs now won't take unlimited space. By default, log files can now take up to 50 MiB of space before a new one is created. Can be changed in the config.

MySQL's service often uses the name "MySQL80" instead of "MySQL", throwing an undesired error when it tries to see if it's running or not. It will now check both names.

Added a new run parameter that will disable the MySQL service checks. This effectively adds a better compatibility with UNIX-like systems.

LoginManager now will try to get the environment variables for Linux, in case it doesn't find the APPDATA variable in the system.